+++ internal/impl/cohere/chat_processor.go
@@ -17,6 +17,7 @@ import (
 
 	cohere "github.com/cohere-ai/cohere-go/v2"
 	"github.com/redpanda-data/benthos/v4/public/service"
+
 	"github.com/redpanda-data/connect/v4/internal/impl/confluent/sr"
 )